Migrate db

+++ db/migrations/20160212095217_CreateTableIncomes.sql
@@ -6,8 +6,8 @@ CREATE TABLE `incomes` (
   `user_id` INT(11) NOT NULL,
   `referer_id` INT(11) NOT NULL,
   `type` TINYINT(4) NOT NULL COMMENT '0: reward, 1: offerwall',
-  `income` MEDIUMINT(8) NOT NULL,
-  `referer_income` MEDIUMINT(8) NOT NULL,
+  `income` DECIMAL(19, 8) NOT NULL,
+  `referer_income` DECIMAL(19, 8) NOT NULL,
   `created_at` DATETIME NOT NULL DEFAULT CURRENT_TIMESTAMP,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
