+++ models/total_reward.go
@@ -5,7 +5,7 @@ import "time"
 // TotalReward model
 type TotalReward struct {
 	ID        int64     `db:"id"`
-	Total     int64     `db:"total"`
+	Total     float64   `db:"total"`
 	CreatedAt time.Time `db:"created_at"`
 }
