+++ services/cache/cache.go
@@ -9,7 +9,7 @@ import (
 // Cache defines interface that one should implement
 type Cache interface {
 	GetLatestTotalReward() models.TotalReward
-	IncrementTotalReward(time.Time, int64)
+	IncrementTotalReward(time.Time, float64)
 
 	GetRewardRatesByType(string) []models.RewardRate
 	SetRewardRates(string, []models.RewardRate)
